refactor: ReservationWaitingInfo 패키지 변경 및 엔티티 hash&equals 추가
